Add retained email address processor
This adds a handler for retained email addresses. Retained emails are
considered to be undeliverable according to Mailingwork due to a
certain number of sequential bounces. Emails on this list are deleted
and a "Contact Updated" activity is created to make note of this
action.

The synchronization performs a full scan on every run by-design in
order to (re-)delete retained addresses should they be re-added by
some other process.