GP-24196 Add error handling
This adds error handling for all processors dealing with mailing-related
entities. When an error is encountered, the sync status is initially set
to "Retrying". Items with this status will be retried once during the
next processor invocation, after which they will be set to "Failed",
effectively stopping further import attempts unless the status is
updated during manual intervention.